♻️ move to config enriching logic into core
